# Avoid allocating temporary views in when decoding strings in `CodedBufferReader`