--- a/protobuf/lib/src/protobuf/coded_buffer_reader.dart
+++ b/protobuf/lib/src/protobuf/coded_buffer_reader.dart
@@ -148,7 +148,14 @@ class CodedBufferReader {
         _buffer.buffer, _buffer.offsetInBytes + _bufferPos - length, length);
   }
 
-  String readString() => _utf8.decode(readBytesAsView());
+  String readString() {
+    final length = readInt32();
+    final stringPos = _bufferPos;
+    _checkLimit(length);
+    return Utf8Decoder(allowMalformed: true)
+        .convert(_buffer, stringPos, stringPos + length);
+  }
+
   double readFloat() => _readByteData(4).getFloat32(0, Endian.little);
   double readDouble() => _readByteData(8).getFloat64(0, Endian.little);
 
@@ -180,7 +187,8 @@ class CodedBufferReader {
         readFixed64();
         return true;
       case WIRETYPE_LENGTH_DELIMITED:
-        readBytesAsView();
+        final length = readInt32();
+        _checkLimit(length);
         return true;
       case WIRETYPE_FIXED32:
         readFixed32();