Benchmark results from the same internal benchmark reported in previous PRs/commits:

|                              | Before     | After      | Diff                |
|------------------------------|------------|------------|---------------------|
| AOT                          |  31,025 us |  27,597 us | - 3,428 us, -11.0%  |
| JIT                          |  34,829 us |  26,846 us | - 7,983 us, -22.9%  |
| dart2js -O4                  | 300,571 us | 248,300 us | -52,271 us, -17.3%  |
| dart2wasm --omit-type-checks | 130,812 us | 107,850 us | -22,962 us, -17.5%  |